backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
langchain
langchain-huggingface
python-jose[cryptography]
passlib[bcrypt]
python-dotenv